Feature/wendy

+++ src/components/mock investment/eachintereststock.js
@@ -4,29 +4,70 @@ import { useNavigate } from "react-router-dom";
 // 관심종목의 각 주식
 export default function Eachintereststock(props) {
   const navigate = useNavigate();
-  // const token = localStorage.getItem("access_token");
-  const token =
-    "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiIzNjc2NTcwMjEyIiwiZXhwIjoxNzI2MDUzNDc2fQ.B9fZwY5dsWvOgPL1TOpILL-4ejYeUMiOoqNxizV51fWS6S_R_hKoLkTEyB_prkKMCtRkNfQD-L62xaMZmwCCdg";
+  const token = localStorage.getItem("access_token");
   // useState를 사용하여 이미지의 경로를 관리하는 상태를 선언합니다.초기값으로 'a.jpg'를 설정합니다.
   const [imageSrc, setImageSrc] = useState("/assets/images/interest.png");
 
   // 이미지 경로를 변경하는 함수를 정의.
-  const toggleImage = (stock_id) => {
-    // 좋아요 되어있는거 클릭해서 없애기
-    fetch("https://heartfolio.site/api/stock/favorites/" + stock_id, {
-      credentials: "include",
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${token}`, // 토큰을 헤더에 추가
-        "Content-Type": "application/json",
-      },
-    }).then((response) => {
+  const toggleImage = async (stock_id) => {
+    let response;
+    try {
+      // 좋아요 되어있는거 클릭해서 없애기
+      response = await fetch(
+        "https://heartfolio.site/api/stock/favorites/" + stock_id,
+        {
+          credentials: "include",
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${token}`, // 토큰을 헤더에 추가
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      if (response.status === 401) {
+        // Access token 만료 -> refresh token으로 새 access token 요청
+        const refreshToken = localStorage.getItem("refresh_token");
+        const refreshResponse = await fetch(
+          "https://heartfolio.site/api/auth/refresh-token",
+          {
+            method: "POST",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ refreshToken: refreshToken }),
+          }
+        );
+
+        if (refreshResponse.status === 200) {
+          const data = await refreshResponse.json();
+          localStorage.setItem("access_token", data.accessToken); // 새 access token 저장
+
+          // 새로운 access token으로 원래 요청 다시 시도
+          response = await fetch(
+            "https://heartfolio.site/api/stock/favorites/" + stock_id,
+            {
+              credentials: "include",
+              method: "DELETE",
+              headers: {
+                Authorization: `Bearer ${localStorage.getItem("access_token")}`, // 새 access token 사용
+                "Content-Type": "application/json",
+              },
+            }
+          );
+        } else {
+          // refresh token도 만료되거나 오류가 있으면 로그인 페이지로 이동
+          localStorage.removeItem("access_token");
+          window.location.href = "/login";
+          return;
+        }
+      }
+
       if (response.ok) {
         navigate("/intereststock");
       }
       // 페이지 새로고침
       window.location.reload();
-    });
+    } catch (error) {
+      console.error("error:", error);
+    }
   };
 
   // 각 주식 페이지 이동
